chore: improve code quality
