+++ Doodle2Gather/Doodle2Gather/Application/Conference/Controllers/ConferenceViewController.swift
@@ -16,6 +16,10 @@ class ConferenceViewController: UIViewController {
     var isVideoOff = false
     var isChatShown = false
 
+    enum Segues {
+        static let toChat = "toChat"
+    }
+
     override func viewDidLoad() {
         super.viewDidLoad()
         videoEngine = AgoraVideoEngine()
@@ -50,7 +54,7 @@ class ConferenceViewController: UIViewController {
 
     // Passes data to the ChatViewController
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if segue.identifier == "toChat" {
+        if segue.identifier == Segues.toChat {
             guard let nav = segue.destination as? UINavigationController else {
                 return
             }
