+++ Doodle2Gather/Doodle2Gather/Application/Conference/Models/Message.swift
@@ -1,0 +1,8 @@
+import MessageKit
+
+struct Message: MessageType {
+    var sender: SenderType
+    var messageId: String
+    var sentDate: Date
+    var kind: MessageKind
+}
