+++ Doodle2Gather/Doodle2Gather/Application/Conference/Models/Sender.swift
@@ -1,0 +1,6 @@
+import MessageKit
+
+struct Sender: SenderType {
+    var senderId: String
+    var displayName: String
+}
